Updated cache buster logic
Removed css and script busters in favour of a uuid/guid cache
buster that is generated when the app starts.

This change means that in the event your NodeBB crashes often,
users will not be able to rely on their browser cache to serve
an unchanged style/js file, but if that's the case, you've
got bigger problems anyway.

re: #3573